Get codec info from PMS info if possible
startCodecsLoading() is now called twice:
1. Before the stream is even opened. getPlaybackInfo() is made to use
the PMS metadata in this case.
2. After the stream has been opened and before a dedoder starts.
getPlaybackInfo() will use the player-analyzed metadata.

I'm hoping this makes all combinations of possible fuckups work. One
case this should certainly fix is playing raw DTS. If no DTS codec was
present yet, opening would fail in libavformat. Downloading the codec
before opening the file should take care of it.

This is chained with display mode switching. Not bothering with making
it run in parallel with mode switching, because I'm not even sure the
simpler current implementation is always correct.