PlayerComponent: avoid that host and web go out of sync
This fixes the problems with playback not starting.

It turns out that if during loading a pause command is accidentally
sent, the host will go into pause mode. Playback will start paused, as
if web specified autoPlay=false. Web doesn't expect this, and implicitly
assumes that playback is _not_ paused. As a result, web is stuck and
can't unpause, because it thinks playback is active.

Another part of the problem is that a pause command is accidentally
sent. This might be a result of key autorepeat triggering too quickly.
If the user hits "enter" in the preplay screen, "enter" might get
repeated after playback/loading logically starts, resulting in the pause
command.

This commit happens to workaround both issues by forcing the pause state
back to the autoPlay parameter specified by web on loading. Any pause
commands sent during loading will essentially be overridden.

We need to fix these things properly, but it will take time.