SettingsComponent: remember the oldest installed version
This uses separate code via QSettings, so the value survives config
version bumps or user deletions. The user has no disadvantages from
this, so the separate settings should not cause any issues.

To make sure the legacy detection mechanism works, we must sure that the
normal config file is absolutely not written before this code is run,
thus the change in saveSettings().