Make direct hardware decoding work on Intel/Linux
First, we need to force Qt to use EGL instead of GLX. This is required
because the direct hw interop mpv uses requires EGL. On the other hand,
always using EGL would break direct hw interop on nVidia GPUs, because
those require GLX. Qt can't just change the backend on the fly. In fact,
it seems that once the underlying plugin is loaded, it's stuck with the
choice. Some elaborate detection is needed. It probably can't be
achieved with Qt in its current state.

Achieve this with a hack: create a small mpv instance, which probes
whether GLX or EGL is better. This is (probably) better than writing
code to create EGL and GLX contects using the raw APIs.

In addition, PMP needs to hand down the X11 Display handle to mpv, as
mpv can't grab it automatically (as it's possible on GLX). Fortunately,
the X11Extras Qt module provides a sane way to do this.