PlayerComponent: cache hwdec-interop debug info
A terribly obscure interaction with libmpv, PMP's debug overlay, and the
Qt scene graph "basic" renderloop causes video to occasionally stutter
very badly as long as the debug overlay is visible.

The issue is that reading the "hwdec-interop" property will internally
acquire a lock, which waits until video rendering is done. If the timing
is bad, it will wait for the frame that is currently being renderer,
essentially resulting in a deadlock. (It triggers once every few
seconds.)

It does not happen with the "threaded" renderloop, because that one
decouples updating the debug overlay and rendering the video. Normally
we don't use the "basic" renderloop, but ANGLE on Windows forces us to.

This commit works this around by querying the property only once (it
doesn't ever change anyway). The workaround will become unnecessary with
a planned change in mpv, but in order not to hold back everything, apply
it now.

(And yes, this use of the C++ "mutable" keyword is actually exactly what
it was invented for.)