RemoteComponent: add apparently missing locking
It accesses m_subscriber, so it definitely should lock m_subscriberLock.

handleSubscription() also acquires the lock (and it is not recursive),
so there is a logical race condition left. However, this shouldn't cause
any real problems, because the code below can deal with a subscriber
being removed during the race condition window.