Load web-client from resources directory instead of the Qt resource system
Now that the web-client is bundled with the binaries in the resources
directory we want to load it from there directly as well.

Since many configuration point the startup URL to
qrc://konvergo/index.html we need to migrate them to a new value.
Instead of putting the full file:// url in this setting I made a
special value called "bundled" instead that will always point to the
latest version of the bundled web-client. This makes any future
migrations unnecessary.

This is implemented in a new function in SettingsComponent as opposed
to just reading the values raw.