Fix a bug where fullscreen setting could be auto-repeated
This seems to be because on OSX entering or exiting fullscreen
interrupts keyboard input so we never get the release signal.

The solution here was to cancel any auto-repeat right before
we issue the fullscreen command. This was a better solution than
the additional delay I added first (which also happened to fix
the issue but was most likely prone to races).

Fixes plexinc/plex-media-player-private#489