If 7.1 is selected, also allow 5.1 and stereo
The player will select the "best" layout out of a ","-separated list.
Make the 7.1 preset include 5.1 and stereo, and the 5.1 preset include
stereo.

This works only as far as the underlying audio APIs allow it. For
example, OSX to our knowledge always enforces a single multichannel
layout, so you can send e.g. only native 7.1 or 5.1 upmixed to 7.1.

This requires reasonably recent deps (libmpv changes).