Fix subtitle stream selection for vobsubs with multiple streams
Until now, we've always assumed that a subtitle file contains exactly 1
subtitle. This is not true with some vobsub files. We have to abandon
the current simplistic way of selecting streams. Instead, check the
stream list as returned by the player directly, and select the wanted
streams explicitly.

Requires a corresponding web-client change to work, but is backwards
compatible.