Update settings on fullscreen change
The OS could kick us out of fullscreen without our interaction, which
means that we'll need to update the fullscreen setting as well (it would
remain stuck in its old, outdated state if we didn't).

To avoid possible recursion and ping-pong issues when writing the
setting, we need to ignore update notifications if we are changing the
setting ourselves. Since Qt does not have a mechanism for selective
signal blocking, we use a simple counter to to ignore settings
notifications (and a slightly less simple RAII wrapper to reset the
counter in an exception-safe way).