Cleanup settings remove/reset behavior
Instead of letting removeValues() and friends just literally wipe out
all settings, remove them only if they're not described in
settings_descriptions.json. If they are described, then merely reset
them to the default.

This also removes some code which reset values to QVariant(), instead of
the default value.

Rename some instances of "remove" to "reset", because that's what it
actually does.

In addition, send change notifications in cases when resetting a value
or a section changes the values (such as changing it back to default).
Entirely removed values are not included in the change notifications -
the assumption is that nothing cares about them anymore.

All these changes should allow web-client to reset any section safely,
even sections that contain host values.