Revert "Use new RPI rendering method"
This partially reverts commit f5edf1e4324c5e982eb7b89e7099680093903954.

It restores the behavior on RPI before that commit. Though the code is a
bit different: the debug.force_vo option can now be set to "opengl-cb"
to test the new rendering code. We hope to use the new rendering code as
soon as the performance issues have been analyzed.

This commit is actually untested on RPI, so I'm hoping for the best.