PlayerComponent: fallback to first stream if stream index can't be found
This fixes selection of the audio stream with some transcodes. The
reason is that web-client will pass the original file's stream indexes
to the host, which is wrong.

In the long run it'd probably be better to pass explicitly whether
transcoding is used, or so. This commit still has some justification
beyond being a hack to make transcoding work again because e.g. the
server information might be outdated, the server somehow determines
stream indexes differently, or other cases when the indeses mismatch.