DisplayManager: fix more refresh rate selection issues
Exact matches should be strictly preferred over inexact ones. There can
be rates which are multiple+close+both, whose weights add up to more
than the exact match weight (75+50+25=150>100). So increase the exact
match weight.

Also, don't compare the float refresh rate values with ==, as small
deviations will make the comparison fail. It's sort of amazing that it
worked at all. In this case, a rate of 23.9761 was not considered equal
to 23.976.

Note that if we have multiple close matches, we still don't pick the
closest rate. I don't think the weighting approach allows much
improvements here without turning it into a clusterfuck. But the
selection is probably fine now.