PlayerComponent: fix AC3/DTS passthrough breakage
The "advanced" setting was accidentally still queried here. Since it's
usually "false" now by default (because it was removed), it'll disable
any kind of audio passthrough (DTS etc.) in the following situations:

- new installations after the setting was removed (1.1.7 and later)
- the user had the "advanced" setting to false previously

It still worked if the user had enabled "Advanced" in older versions,
and then upgraded.

Manually editing plexmediaplayer.conf and adding

"advanced": true,

to the audio setting works this bug around as well.