Input: better stringification for key input with no text representation
This is mainly for special keys, for which Qt's QKeySequence will
essentially produce garbage text.

Try to avoid such strings, and produce a numeric key code instead. This
should not affect normal keys, for which we (and users) have hardcoded
key names in input maps.

Using QKeyEvent::nativeVirtualKey() is probably the best way to get
something numeric and that isn't completely arbitrary. But it can return
0, in which case we fall back to hex dumping the Qt potential-garbage
string. The suffix to the numbers ("V" and "Q") is supposed to
disambiguate and distinguish them.