OSX: restore moude hiding hack on playback start
This is most likely the VTDecompressionSessionCreate issue again (it
doesn't reproduce if hardware decoding is disabled). We had something
quite similar before, but removed it recently. This commit adds it back,
in a slightly "better" way.

Reproduction steps:
- move mouse cursor to dock
- start PMP (in fullscreen mode)
- use keyboard to start playback (hw decoding on)
- mouse cursor will become visible and stay visible

The bad part is that it "flickers" slightly on start. I'm hoping nobody
minds.