KonvergoWindow: remove a disruptive loadGeometry() call
This is called when the window goes to windowed mode - usually by
itself, because there's another loadGeometry() call in
KonvergoWindow::updateWindowState(), which sets the geometry when
explicitly switching to windowed mode from the PMP side.

I can't find any rason why this call exists. On the other hand, it can
have weird effects when the fullscreen state is lost due to display
hotplugging on Windows. Then the window will have no titlebar, and can
cover the desktop partially. (This is not the same as a similar bug,
which also looks like it covers the desktop partially, but actually it
fully covers the desktop and is merely transparent in some regions.)