SettingsComponent: fix forcing display mode on command line
Reproduction:
- start PMP, set it to fullscreened TV mode, exit
- now start with "--desktop --fullscreen"
- it will start in TV mode

The problem is that the "webMode" setting is never set to "desktop".
Indeed, the "webMode" setting is set from "layout" in this code path,
which includes only "tv" and "auto" anyway. A forced "desktop" layout
mode makes strangely no sense, because there'd be no way for the user
to change it (since desktop mode still lacks PMP settings).

Fix by changing the "webMode" setting to "desktop" explicitly in this
case.