SettingsComponent: add set_setting command
Mostly useful for debugging.

Parsing the settings values by concatenating them with some other
strings to make them a valid JSON document is not really clean,
but avoids tons of problems (and we can have simpler code). In
particular, QVariant::convert() turned out being a POS from what
I can tell.

Syntax:

   host:set_setting <qualified_name> <value>

where <qualified_name> is the settings section and value name
separated by a ".", e.g. 'main.fullscreen' (without the quotes),
and <value> is a JSON value. For strings, this would be '"str"'
(without the outer quotes). An example for an input map would be:

   "F1": "host:set_setting main.foo \"bar\"",

which sets the "main.foo" setting to the string "bar" (without
the quotes). The quotes have to be quoted, because they're
already inside of a string. For numbers or bools, no quoting
is required.