KonvergoWindow: add a hidden forceAlwaysFS setting
If set to true, it aggressively tries to enter or re-enter fullscreen
mode. It's probably too aggressive and obscure as that I'd want it to be
visible by default.

This will probably help with Windows fullscreen&hotplugging problems,
although since they are so hard to reproduce, I didn't get to test it.