webview.qml: swap F11 and Shift+F11 in desktop mode for consistency
It appears this was a mistake. Now the shortcuts are consistent with
input.json, and with the shortcuts for MacOS.

Shift+F11 will always toggle fullscreen and not switch the mode (both in
TV and desktop mode). F11 will always toggle fullscreen and web mode,
like the UI button does.