KonvergoWindow: improve web scale factor notification handling
Strictly call it if any of those change. I'm not entirely sure how their
relation is, but maybe it's good if the code doesn't have to care about
it. Also, cache the result, instead of possibly computing inconsistent
in-between values of it (e.g. during resizing). This makes sense
especially because notifyScale() gets a special size value as argument,
that apparently doesn't have to be the same as the window's size().