KonvergoWindow: make sure newly loaded web-client has correct initial-scale
Do this by recomputing the correct web scale before the URL change event
is emitted.

This probably fixes some cases of web-client not resizing properly when
changing web mode _and_ fullscreen mode.

It doesn't always work, so the merit is unknown.

Also, the other updateWindowState() and notifyScale() calls are not
needed.