SystemComponent: expose web scale as explicit property
Web-client needs the correct scale at initialization. Until now, we've
attempted to do this via the "initialScale" URL parameter. But that
leaves a time window where the scale can change, but the web-client has
not setup a change listener yet. The solution is makign web-client read
the property on initialization.

The "initialScale" thing can probably be removed later. Not sure.

This actually fixes the fullscreen/web mode switching scale race
condition. Requires web-client update for that to work (which is not
doine yet).