Enable refreshrate.avoid_25hz_30hz by default
There's no real reason to prefer 25hz over 50hz, or 30hz over 60hz,
ever. One might argue that motion interpolation could benefit from this,
but that's a bit pointless if the refresh rate is an integer multiple of
the video rate. If you really want that, just disable the option.