installer: do not remove shortcuts when updating, fixes #425
Current installer will remove completely the previous installed version before installing the update
This results in breaking the pinned taskbar shortcuts at every update which is irritating for users.

That patch will skip the shortcuts removal. This will also skip the shortcut removal when completely uninstalling the app, but windows will propose to remove it next time you click it and that the app was uninstalled.

So this is improving current situation even if not perfect.