osx: remove presentation options
those options have been added long ago to avoid having troubles with topbar that would remain in some cases. Now this issue seems fixed.

It also has a side effect with OSX not allowing to show the topbar as other apps would when you hover the mouse on it.

Removing this to restore some OSX "standard behavior".