Properly enable QWebEngine GPU acceleration
Mentioned here: https://github.com/jellyfin/jellyfin-media-player/issues/10

I noticed the interface was lagging a lot when using a theme that has blurred backdrops (disabling the backdrops made it smooth again so I know that was the culprit). I knew it was because of a lack of hardware acceleration, and I stumbled across this issue and realized it was never actually enabled.

The difference in performance is night and day for me when compiled with this flag enabled.