Add build instructions for Fedora systems.
Added build instructions that I wrote in #343. I have been testing this over the past couple months and they have been working for me. Added instructions at the beginning of the build commands for Linux on downloading the latest release tag instead of just pulling the main branch as this should make for more consistent builds.